# api_gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.LearningSkill:
    properties:
      created_at:
        type: string
      proficiency_level:
        type: string
      skill_id:
        type: string
      skill_name:
        type: string
      user_id:
        type: string
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.LearningSkillUpdate:
    properties:
      proficiency_level:
        type: string
      skill_name:
        type: string
    required:
    - proficiency_level
    - skill_name
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.LearningSkillsUpdate:
    properties:
      learning_skills:
        items:
          $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.LearningSkillUpdate'
        type: array
    type: object
...
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.MentorProfile:
    properties:
      created_at:
        type: string
      description:
        type: string
      rating:
        type: number
      user_id:
        type: string
      withdrawal_address:
        type: string
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.MentorUpdate:
    properties:
      description:
        type: string
      withdrawal_address:
        type: string
    type: object
...
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ProfileResponse:
    properties:
      learning_skills:
        items:
          $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.LearningSkill'
        type: array
      mentor:
        $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.MentorProfile'
      student:
        $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.StudentProfile'
      teaching_skills:
        items:
          $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.TeachingSkill'
        type: array
      user:
        $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.User'
    type: object
...
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.StudentProfile:
    properties:
      created_at:
        type: string
      learning_goals:
        type: string
      preferred_learning_style:
        type: string
      user_id:
        type: string
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.StudentUpdate:
    properties:
      learning_goals:
        type: string
      preferred_learning_style:
        type: string
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.TeachingSkill:
    properties:
      created_at:
        type: string
      proficiency_level:
        type: string
      skill_id:
        type: string
      skill_name:
        type: string
      user_id:
        type: string
      years_of_experience:
        type: integer
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.TeachingSkillUpdate:
    properties:
      proficiency_level:
        type: string
      skill_name:
        type: string
      years_of_experience:
        type: integer
    required:
    - proficiency_level
    - skill_name
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.TeachingSkillsUpdate:
    properties:
      teaching_skills:
        items:
          $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.TeachingSkillUpdate'
        type: array
    type: object
...
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      learning_skills:
        $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.LearningSkillsUpdate'
      mentor_data:
        $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.MentorUpdate'
      student_data:
        $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.StudentUpdate'
      teaching_skills:
        $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.TeachingSkillsUpdate'
    type: object
...
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      user_id:
        type: string
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.UserCountResponse:
    properties:
      user_count:
        type: string
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.SuccessResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Gateway для Mentor Platform
  title: Mentor Platform API
  version: "1.0"
paths:
...
  /profiles/{id}:
    get:
      description: Возвращает полную информацию о профиле пользователя включая менторские/студенческие
        данные
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ProfileResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
      summary: Получить профиль
      tags:
      - profiles
    put:
      consumes:
      - application/json
      description: Обновляет данные профиля пользователя. Автоматически создает mentor/student
        записи при необходимости.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить профиль
      tags:
      - profiles
...
  /users/{id}:
    delete:
      description: Удаляет пользователя и все связанные данные
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить пользователя
      tags:
      - users
    get:
      description: Возвращает информацию о пользователе по его идентификатору
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
      summary: Получить пользователя по ID
      tags:
      - users
  /users/all:
    get:
      description: Возвращает количество пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.UserCountResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
      summary: Получить количество пользователей сайта
      tags:
      - users
  /users/email/{email}:
    get:
      description: Возвращает информацию о пользователе по его email
      parameters:
      - description: Email пользователя
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_utils.ErrorResponse'
      summary: Получить пользователя по email
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Введите: Bearer {jwt_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"